pyproject.toml
[tool.poetry]
name = "amongus_displace"
version = "0.1.0"
description = ""
authors = ["Santiago Cézar <santiagocezar2013@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
Pillow = "^9.1.0"
numpy = "^1.22.3"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
black = {version = "^22.3.0", allow-prereleases = true}

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
